More work and some restructuring in morphing code

+++ madminer/tools/h5_interface.py
@@ -113,7 +113,7 @@ def load_madminer_settings(filename):
         # Morphing
         try:
             morphing_components = np.asarray(f['morphing/components'][()], dtype=np.int)
-            morphing_matrix = np.asarray(f['morphing/components'][()], dtype=np.int)
+            morphing_matrix = np.asarray(f['morphing/components'][()])
 
         except:
             morphing_components = None
